rename file and consolidate similar files
